first cut check in.
