+++ docker-compose.selenium.yml
@@ -44,11 +44,11 @@ services:
     ports:
       - "8080:8080"
       - "7890:7890"
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://msls:8080/sso/authorize"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
+    # healthcheck:
+    #   test: ["CMD", "curl", "-f", "http://msls:8080/sso/authorize"]
+    #   interval: 10s
+    #   timeout: 5s
+    #   retries: 5
 
   db:
     image: postgres
@@ -84,6 +84,6 @@ services:
       timeout: 20s
     depends_on:
       msls:
-        condition: service_healthy
+        condition: service_started
       db:
         condition: service_healthy
